% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/givd_summary.R
\name{givd_summary}
\alias{givd_summary}
\title{GIVD Metadata}
\usage{
givd_summary(obj)
}
\arguments{
\item{obj}{A vegtable object.}
}
\description{
Get metadata for the GIVD mask.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
