% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_delete_concept.R
\name{pg_delete_concept}
\alias{pg_delete_concept}
\alias{pg_delete_name}
\title{Delete Names or Concepts in PostgreSQL Taxonomic Lists}
\usage{
pg_delete_concept(conn, schema, concept_id)

pg_delete_name(conn, schema, usage_id)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{schema}{Name of the schema containing taxonomic tables.}

\item{concept_id, usage_id}{ID of taxon concept or taxon usage name to be
deleted.}
}
\description{
Deletion of synonyms or taxon concepts in a PostgreSQL version of
\linkS4class{taxlist} objects.

Take care of producing backups before starting the manipulation of
databases.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
