% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_merge_names.R
\name{pg_merge_names}
\alias{pg_merge_names}
\alias{pg_merge_names.PostgreSQLConnection}
\title{Merge names across data sets}
\usage{
pg_merge_names(conn, ...)

\method{pg_merge_names}{PostgreSQLConnection}(
  conn,
  new_id,
  old_id,
  taxon_names = c("tax_commons", "taxonNames"),
  schemas = c("swea_dataveg", "sudamerica"),
  ...
)
}
\arguments{
\item{conn}{A database connection established by \code{\link[=dbConnect]{dbConnect()}}.}

\item{...}{Further arguments passed among methods.}

\item{new_id}{Integer value, the persisting name id ('TaxonUsageID').}

\item{old_id}{Integer value, the name id ('TaxonUsageID') to be replaced.}

\item{taxon_names}{Character vector indicating the name of schema and table
containing taxon names.}

\item{schemas}{Character vector indicating the schemas containing the
respective databases sharing taxon names.}
}
\description{
When completing authors names in common names tables, duplicated entries may
emerge.
To merge incomplete entry with a complete one, the replacement have to be
done across data sets (vegetation-plot databases).
}
