% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pg_replace_community.R
\name{pg_replace_community}
\alias{pg_replace_community}
\title{Replace communities within veg_databases}
\usage{
pg_replace_community(conn, old, new)
}
\arguments{
\item{conn}{A connection established by \code{\link[=dbConnect]{dbConnect()}}.}

\item{old}{An integer value with the community ID (community_type) that will
be replaced and deleted at the end.}

\item{new}{An integer value with the community ID that will be inserted in
exchange to 'old'.}
}
\description{
This function replaces entries of 'old' by 'new' across all stored databases.
The community declared at 'old' will be deleted from the main table at the
end of the replacement.
}
