% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_communities.R
\name{report_communities}
\alias{report_communities}
\alias{report_communities,vegtable-method}
\title{Summary on plant communities stored in vegtable objects}
\usage{
report_communities(veg, ...)

\S4method{report_communities}{vegtable}(
  veg,
  bib,
  filename,
  title,
  author,
  date,
  date_format = "\%d-\%m-\%Y",
  papersize = "a4",
  keep_rmd = TRUE,
  biblio_style = "abbrvnat",
  ...
)
}
\arguments{
\item{veg}{An object of class \linkS4class{vegtable}.}

\item{...}{Further arguments passed among methods.}

\item{bib}{Name of bibliography file (.bib) as chararcter value.}

\item{filename}{A character value indicating the name of the resulting file
without extension.}

\item{title, author}{Character values used in the title page.}

\item{date}{A value of class \code{\link{Date}} to be inserted in the title
page.}

\item{date_format}{Character value indicating the format used for displaying
the date in the title page.}

\item{papersize}{Character value indicating the paper format.}

\item{keep_rmd}{Logical value, whether the resulting Rmd file should be
deleted after rendering or not.}

\item{biblio_style}{Character value indicating the name of the bibliography
style in the package 'natbib'.}
}
\value{
A PDF file.
}
\description{
This function produces a report on plant communities stored in an object of
class \linkS4class{vegtable}.
The function is adapted to our own databases and may not work in other data
sets.

This function generates a quick overview of references and communities
stored in a data set formatted as \linkS4class{vegtable} object.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
