% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxlist2tvlist.R
\name{taxlist2tvlist}
\alias{taxlist2tvlist}
\title{Convert taxlist objects to Turboveg format}
\usage{
taxlist2tvlist(taxlist, ecodbase = TRUE)
}
\arguments{
\item{taxlist}{A \linkS4class{taxlist} object.}

\item{ecodbase}{A logical value, whether the slot \code{'taxonTraits'} may
be exported or not.}
}
\description{
This is a function converting \linkS4class{taxlist} objects to a \strong{Turboveg} like
format.
It may be used as prototype for an eventual export function.

This function basically export the slots \code{taxonNames} and \code{taxonTraits},
adapts the respective column names to \strong{Turboveg}, and generates
automatically some fields that are usually not requested in \linkS4class{taxlist}
objects.

Names belonging to a same concept get a common \code{LETTERCODE}.
}
\seealso{
\linkS4class{taxlist} \code{\link[=tv2taxlist]{tv2taxlist()}}
}
\author{
Miguel Alvarez \email{malvarez@uni-bonn.de}
}
