\name{version_svg}
\alias{version_svg}

\title{Draw a SVG badge for database version}
\description{
Function for convenience.
}
\usage{
version_svg(file, version)
}
\arguments{
\item{file}{A character value with the name of the badge without extension.}
\item{version}{A character value indicating the version of the database.}
}
\details{
As version the date formatted accordingly but also another text is possible.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}.
}
