% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postgres2vegtable.R
\name{postgres2vegtable}
\alias{postgres2vegtable}
\alias{import_sudamerica}
\alias{import_swea}
\alias{import_bernice}
\title{Import PostgreSQL databases into vegtable objects}
\usage{
postgres2vegtable(
  conn,
  tax_args = list(),
  header,
  sql_header,
  samples,
  relations = list(),
  layers = list(),
  coverconvert = list(),
  geometry,
  description,
  as_list = FALSE,
  ...
)

import_sudamerica(
  conn,
  header = c("sudamerica", "header"),
  samples = c("sudamerica", "samples"),
  relations = list(community_type = c("commons", "community_type"), data_source =
    c("commons", "data_source")),
  layers = list(spec_miguel = c("specimens", "specimens_miguel")),
  coverconvert = list(br_bl = c("commons", "br_bl"), b_bbds = c("commons", "b_bbds"),
    ordinal = c("commons", "ordinal")),
  geometry = "plot_centroid",
  get_countries = TRUE,
  ...
)

import_swea(
  conn,
  header = c("swea_dataveg", "header"),
  samples = c("swea_dataveg", "samples"),
  relations = list(globe_plots = c("swea_dataveg", "globe_plots"), swea1_code =
    c("swea_dataveg", "swea1_code"), soil_moisture = c("swea_dataveg", "soil_moisture"),
    soil_texture = c("swea_dataveg", "soil_texture"), community_type = c("commons",
    "community_type"), data_source = c("commons", "data_source"), naturalness =
    c("swea_dataveg", "naturalness"), record_type = c("swea_dataveg", "record_type")),
  layers = list(veg_layer = c("swea_dataveg", "veg_layer"), spec_miguel =
    c("specimens", "specimens_miguel")),
  coverconvert = list(br_bl = c("commons", "br_bl"), b_bbds = c("commons", "b_bbds"),
    ordinal = c("commons", "ordinal")),
  geometry = "plot_centroid",
  get_countries = TRUE,
  ...
)

import_bernice(
  conn,
  description = c(Title = "Parthenium survey in Lake Baringo", Author =
    "Bernice Mereina Sainepo", Source = "SWEA-Dataveg (GIVD-AF-00-006)", Version =
    Sys.Date()),
  head_cols = c("ReleveID", "code_trr228", "original_number", "record_date",
    "plot_size", "data_source", "elevation"),
  samples_cols = c("record_id", "ReleveID", "quadrant", "TaxonUsageID",
    "misspelled_name", "cover_percentage", "frequency"),
  ...
)
}
\arguments{
\item{conn}{A database connection provided by \code{\link[=dbConnect]{dbConnect()}}.}

\item{tax_args}{List of arguments passed to \code{\link[=postgres2taxlist]{postgres2taxlist()}}.}

\item{header, samples}{Character vectors indicating the schema and table
containing header and samples information, respectively.}

\item{sql_header}{SQL statement to be used instead of \code{header}.}

\item{relations, layers, coverconvert}{Lists of vectors for the respective
slots, each containing schema and name of required table.}

\item{geometry}{Name of the variable in header containing the geometry of
the plots.}

\item{description}{Named vector with metadata.}

\item{as_list}{Logical value indicating whether a list or an object of class
\linkS4class{vegtable} should be returned.}

\item{...}{Further arguments passed among methods.}

\item{get_countries}{Logical argument, specific for the databases
'sudamerica' and 'SWEA-Dataveg', indicating whether country information
should be reimported from integrated map.}

\item{head_cols}{Character vector indicating the header variables to be
imported (except the coordinates).}

\item{samples_cols}{Character vector indicating the samples variables to be
imported.}
}
\description{
Import and adaption of Postgres tables into objects of class
\linkS4class{vegtable}.

In the case that some schemas are not mentioned, the function assumes such
tables are located in the same schema as the table header. Thus for
databases placed in just one schema, this need to be set only in argument
\code{header_schema}.
}
\author{
Miguel Alvarez, \email{kamapu78@gmail.com}.
}
