% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/version_svg.R
\name{version_svg}
\alias{version_svg}
\title{Draw a SVG badge for database version}
\usage{
version_svg(file, version)
}
\arguments{
\item{file}{A character value with the name of the badge without extension.}

\item{version}{A character value indicating the version of the database.}
}
\description{
Function for convenience.

As version the date formatted accordingly but also another text is possible.
}
\author{
Miguel Alvarez \email{kamapu78@gmail.com}
}
